requirements.txt
PyQt5>=5.15.0
pycryptodome>=3.14.1
androguard>=3.4.0
pyyaml>=6.0
loguru>=0.6.0
